+++ src/Libraries/SmartStore.Services/Tasks/ScheduleTaskService.cs
@@ -83,7 +83,7 @@ namespace SmartStore.Services.Tasks
             {
                 query = query.Where(t => t.Enabled);
             }
-            query = query.OrderByDescending(t => t.Seconds);
+            query = query.OrderBy(t => t.Seconds);
 
             var tasks = query.ToList();
             return tasks;