+++ src/Presentation/SmartStore.Web/Administration/Controllers/ScheduleTaskController.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Web.Mvc;
 using System.Threading.Tasks;
+using System.Reflection;
 using SmartStore.Admin.Models.Directory;
 using SmartStore.Admin.Models.Tasks;
 using SmartStore.Core.Domain.Directory;
@@ -16,6 +17,10 @@ using SmartStore.Services.Tasks;
 using SmartStore.Web.Framework;
 using SmartStore.Web.Framework.Controllers;
 using Telerik.Web.Mvc;
+using SmartStore.Core.Async;
+using Autofac;
+using SmartStore.Core.Logging;
+using SmartStore.Core.Plugins;
 
 namespace SmartStore.Admin.Controllers
 {
@@ -46,6 +51,16 @@ namespace SmartStore.Admin.Controllers
 
         #region Utility
 
+		private bool IsTaskInstalled(ScheduleTask task)
+		{
+			var type = Type.GetType(task.Type);
+			if (type != null)
+			{
+				return PluginManager.IsActivePluginAssembly(type.Assembly);
+			}
+			return false;
+		}
+
         [NonAction]
         protected ScheduleTaskModel PrepareScheduleTaskModel(ScheduleTask task)
         {
@@ -59,7 +74,18 @@ namespace SmartStore.Admin.Controllers
                 LastStartUtc = task.LastStartUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastStartUtc.Value, DateTimeKind.Utc).ToString("G") : "",
                 LastEndUtc = task.LastEndUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastEndUtc.Value, DateTimeKind.Utc).ToString("G") : "",
                 LastSuccessUtc = task.LastSuccessUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastSuccessUtc.Value, DateTimeKind.Utc).ToString("G") : "",
+				LastError = task.LastError.EmptyNull(),
+				IsRunning = task.LastStartUtc.GetValueOrDefault() > task.LastEndUtc.GetValueOrDefault(),
+				Duration = ""
             };
+
+			var span = TimeSpan.Zero;
+			if (task.LastStartUtc.HasValue)
+			{
+				span = model.IsRunning ? DateTime.UtcNow - task.LastStartUtc.Value : task.LastEndUtc.Value - task.LastStartUtc.Value;
+				model.Duration = span.ToString("g");
+			}
+
             return model;
         }
 
@@ -77,15 +103,7 @@ namespace SmartStore.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageScheduleTasks))
                 return AccessDeniedView();
 
-            var models = _scheduleTaskService.GetAllTasks(true)
-                .Select(PrepareScheduleTaskModel)
-                .ToList();
-            var model = new GridModel<ScheduleTaskModel>
-            {
-                Data = models,
-                Total = models.Count
-            };
-            return View(model);
+            return View();
         }
 
         [HttpPost, GridAction(EnableCustomBinding = true)]
@@ -93,8 +111,9 @@ namespace SmartStore.Admin.Controllers
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageScheduleTasks))
                 return AccessDeniedView();
-
+			
             var models = _scheduleTaskService.GetAllTasks(true)
+				.Where(IsTaskInstalled)
                 .Select(PrepareScheduleTaskModel)
                 .ToList();
             var model = new GridModel<ScheduleTaskModel>
@@ -134,6 +153,55 @@ namespace SmartStore.Admin.Controllers
 
             return List(command);
         }
+
+		public ActionResult RunJob(int id, string returnUrl = "")
+		{
+			if (!_permissionService.Authorize(StandardPermissionProvider.ManageScheduleTasks))
+				return AccessDeniedView();
+
+			returnUrl = returnUrl.NullEmpty() ?? Request.UrlReferrer.ToString();
+
+			var t = AsyncRunner.Run(c =>
+			{
+				try
+				{
+					var svc = c.Resolve<IScheduleTaskService>();
+
+					var scheduleTask = svc.GetTaskById(id);
+					if (scheduleTask == null)
+						throw new Exception("Schedule task cannot be loaded");
+
+					var job = new Job(scheduleTask);
+					job.Enabled = true;
+					job.Execute(c, false);
+				}
+				catch
+				{
+				}
+
+			});
+
+			// wait only 100 ms.
+			t.Wait(100);
+
+			if (t.IsCompleted)
+			{
+				if (!t.IsFaulted)
+				{
+					NotifySuccess(T("Admin.System.ScheduleTasks.RunNow.Completed"));
+				}
+				else
+				{
+					NotifyError(t.Exception.Flatten().InnerException);
+				}
+			}
+			else
+			{
+				NotifyInfo(T("Admin.System.ScheduleTasks.RunNow.Progress"));
+			}
+
+			return Redirect(returnUrl);
+		}
 
         #endregion
     }