+++ src/Presentation/SmartStore.Web/Administration/Models/Tasks/ScheduleTaskModel.cs
@@ -30,5 +30,12 @@ namespace SmartStore.Admin.Models.Tasks
 
         [SmartResourceDisplayName("Admin.System.ScheduleTasks.LastSuccess")]
         public string LastSuccessUtc { get; set; }
+
+		[SmartResourceDisplayName("Common.Error")]
+		public string LastError { get; set; }
+
+		public string Duration { get; set; }
+
+		public bool IsRunning { get; set; }
     }
 }